refactor: seperate parser and validator
