+++ src/test/java/util/ParserTest.java
@@ -1,45 +1,22 @@
 package util;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.Assertions.assertThat;
 
+import java.util.List;
 import org.junit.jupiter.api.Test;
 
 public class ParserTest {
-    public static final String ERROR_PREFIX = "[ERROR] ";
+    Parser parser = new Parser();
 
     @Test
-    void 빈_참가자_이름_예외_테스트() {
+    void 참가자_이름_정상_파싱_테스트() {
         // given
-        String participantsName = "";
+        String request = "영기,라이";
 
-        // when & then
-        assertThatThrownBy(() ->
-                Parser.validateEmptyInput(participantsName))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageStartingWith(ERROR_PREFIX);
-    }
+        // when
+        List<String> response = parser.parseParticipantsName(request);
 
-    @Test
-    void 참가자_이름_숫자_예외_테스트() {
-        // given
-        String participantsName = "123";
-
-        // when & then
-        assertThatThrownBy(() ->
-                Parser.validateNonLiteralInput(participantsName))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageStartingWith(ERROR_PREFIX);
-    }
-
-    @Test
-    void 쉼표_외의_특수문자_예외_테스트() {
-        // given
-        String participantsName = "영기:라이";
-
-        // when & then
-        assertThatThrownBy(() ->
-                Parser.validateInvalidSymbolInput(participantsName))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageStartingWith(ERROR_PREFIX);
+        // then
+        assertThat(response).isEqualTo(List.of("영기", "라이"));
     }
 }
