+++ src/test/java/util/ValidatorTest.java
@@ -14,7 +14,7 @@ public class ValidatorTest {
 
         // when & then
         assertThatThrownBy(() ->
-                Validator.validateEmptyInput(answer))
+                Validator.validateParticipantEmptyInput(answer))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageStartingWith(ERROR_PREFIX);
     }
@@ -27,6 +27,42 @@ public class ValidatorTest {
         // when & then
         assertThatThrownBy(() ->
                 Validator.validateYesOrNo(answer))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageStartingWith(ERROR_PREFIX);
+    }
+
+    @Test
+    void 빈_참가자_이름_예외_테스트() {
+        // given
+        String participantsName = "";
+
+        // when & then
+        assertThatThrownBy(() ->
+                Validator.validateParticipantEmptyInput(participantsName))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageStartingWith(ERROR_PREFIX);
+    }
+
+    @Test
+    void 참가자_이름_숫자_예외_테스트() {
+        // given
+        String participantsName = "123";
+
+        // when & then
+        assertThatThrownBy(() ->
+                Validator.validateNonLiteralInput(participantsName))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageStartingWith(ERROR_PREFIX);
+    }
+
+    @Test
+    void 쉼표_외의_특수문자_예외_테스트() {
+        // given
+        String participantsName = "영기:라이";
+
+        // when & then
+        assertThatThrownBy(() ->
+                Validator.validateInvalidSymbolInput(participantsName))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageStartingWith(ERROR_PREFIX);
     }
